[feature] create a controller for gemini screening and connect it to frontend
