+++ backend/src/routes/results.routes.ts
@@ -1,10 +1,17 @@
 import { Router } from "express";
+import multer from "multer";
+import * as geminiScreening from "../controllers/geminiScreening.controller.js";
 import * as results from "../controllers/results.controller.js";
 
 const router = Router();
+const upload = multer({
+  storage: multer.memoryStorage(),
+  limits: { fileSize: 10 * 1024 * 1024 },
+});
 
 router.get("/", results.list);
 router.post("/", results.create);
+router.post("/screenings/run", upload.single("jobDescriptionFile"), geminiScreening.runScreeningWithGemini);
 router.post("/screenings", results.createScreeningRun);
 router.get("/history", results.getScreeningHistory);
 router.get("/job_posts/:jobPostId", results.getScreeningByJobPostId);
